refactor(json): tighten writer responsibility boundaries
Why:
- ultra-review identified that index writer helpers and CLI writer fallbacks weakened the dedicated writer boundaries for public JSON output.

What Changed:
- move index ordering policy under the index ordering responsibility and keep the shared index writer base to generic array writing only.
- move concrete index entry writers into their owning contract writers.
- require CLI commands to receive ICommandResultWriter through constructor injection and update command tests with a shared test writer factory.

Impact:
- public JSON field names and protocol output remain unchanged while writer ownership and dependency injection are stricter.
