+++ src/Ucli.Contracts/Index/Catalogs/Schemas/IndexSchemasCatalogJsonContractWriter.cs
@@ -22,4 +22,36 @@ internal sealed class IndexSchemasCatalogJsonContractWriter : IndexJsonContractW
     {
         return entries == null ? null : IndexJsonOrderingPolicy.OrderSchemaEntries(entries);
     }
+
+    private static void WriteSchemaEntry (
+        Utf8JsonWriter writer,
+        IndexSchemaEntryJsonContract entry)
+    {
+        writer.WriteStartObject();
+        WriteNullableString(writer, "schemaKey", entry.SchemaKey);
+        WriteNullableString(writer, "kind", entry.Kind);
+        WriteNullableString(writer, "typeId", entry.TypeId);
+        WriteNullableString(writer, "displayName", entry.DisplayName);
+        WriteArray(writer, "properties", OrderSchemaPropertiesOrNull(entry.Properties), WriteSchemaProperty);
+        writer.WriteEndObject();
+    }
+
+    private static void WriteSchemaProperty (
+        Utf8JsonWriter writer,
+        IndexSchemaPropertyEntryJsonContract property)
+    {
+        writer.WriteStartObject();
+        WriteNullableString(writer, "path", property.Path);
+        WriteNullableString(writer, "propertyType", property.PropertyType);
+        WriteNullableString(writer, "declaredTypeId", property.DeclaredTypeId);
+        writer.WriteBoolean("isArray", property.IsArray);
+        WriteNullableString(writer, "elementTypeId", property.ElementTypeId);
+        writer.WriteBoolean("isReadOnly", property.IsReadOnly);
+        writer.WriteEndObject();
+    }
+
+    private static IReadOnlyList<IndexSchemaPropertyEntryJsonContract>? OrderSchemaPropertiesOrNull (IReadOnlyList<IndexSchemaPropertyEntryJsonContract>? properties)
+    {
+        return properties == null ? null : IndexJsonOrderingPolicy.OrderSchemaProperties(properties);
+    }
 }
