+++ src/Ucli/Hosting/Cli/Requests/Query/QueryAssetSchemaCommand.cs
@@ -21,10 +21,10 @@ internal sealed class QueryAssetSchemaCommand
     /// <param name="commandResultWriter"> The command-result writer dependency. </param>
     public QueryAssetSchemaCommand (
         IQueryService queryService,
-        ICommandResultWriter? commandResultWriter = null)
+        ICommandResultWriter commandResultWriter)
     {
         this.queryService = queryService ?? throw new ArgumentNullException(nameof(queryService));
-        this.commandResultWriter = commandResultWriter ?? CommandResultWriter.CreateDefault();
+        this.commandResultWriter = commandResultWriter ?? throw new ArgumentNullException(nameof(commandResultWriter));
     }
 
     /// <summary> Executes <c>query asset schema</c> and emits the JSON result contract. </summary>
