+++ src/Ucli/Hosting/Cli/Requests/Query/QueryAssetsFindCommand.cs
@@ -20,10 +20,10 @@ internal sealed class QueryAssetsFindCommand
     /// <param name="commandResultWriter"> The command-result writer dependency. </param>
     public QueryAssetsFindCommand (
         IQueryService queryService,
-        ICommandResultWriter? commandResultWriter = null)
+        ICommandResultWriter commandResultWriter)
     {
         this.queryService = queryService ?? throw new ArgumentNullException(nameof(queryService));
-        this.commandResultWriter = commandResultWriter ?? CommandResultWriter.CreateDefault();
+        this.commandResultWriter = commandResultWriter ?? throw new ArgumentNullException(nameof(commandResultWriter));
     }
 
     /// <summary> Executes <c>query assets find</c> and emits the JSON result contract. </summary>
