+++ tests/Ucli.Tests/Hosting/Cli/PlanCommandTests.cs
@@ -41,7 +41,7 @@ public sealed class PlanCommandTests
                 PlanToken: "plan-token-1"),
             "uCLI plan completed.")));
         var preflightService = new StubPlanCommandPreflightService((_, _, _, _) => throw new InvalidOperationException("Preflight should not be called."));
-        var command = new PlanCommand(service, preflightService, new StubRequestInputReader(RequestInputReadResult.Success(DefaultRequestJson)));
+        var command = new PlanCommand(service, preflightService, new StubRequestInputReader(RequestInputReadResult.Success(DefaultRequestJson)), CommandResultTestWriter.Create());
         using var cancellationTokenSource = new CancellationTokenSource();
 
         var (exitCode, standardOutput) = await StandardOutputCapture.Execute(() => command.Plan(
@@ -103,7 +103,7 @@ public sealed class PlanCommandTests
                     fallbackReason: null),
                 PlanToken: null))));
         var preflightService = new StubPlanCommandPreflightService((_, _, _, _) => throw new InvalidOperationException("Preflight should not be called."));
-        var command = new PlanCommand(service, preflightService, new StubRequestInputReader(RequestInputReadResult.Success(DefaultRequestJson)));
+        var command = new PlanCommand(service, preflightService, new StubRequestInputReader(RequestInputReadResult.Success(DefaultRequestJson)), CommandResultTestWriter.Create());
 
         var (exitCode, standardOutput) = await StandardOutputCapture.Execute(() => command.Plan(
             projectPath: "/repo/UnityProject",
@@ -126,7 +126,7 @@ public sealed class PlanCommandTests
     {
         var service = new StubPlanService((_, _) => throw new InvalidOperationException("Service should not be called."));
         var preflightService = new StubPlanCommandPreflightService((_, _, _, _) => throw new InvalidOperationException("Preflight should not be called."));
-        var command = new PlanCommand(service, preflightService, new StubRequestInputReader(RequestInputReadResult.Success(DefaultRequestJson)));
+        var command = new PlanCommand(service, preflightService, new StubRequestInputReader(RequestInputReadResult.Success(DefaultRequestJson)), CommandResultTestWriter.Create());
 
         var (exitCode, standardOutput) = await StandardOutputCapture.Execute(() => command.Plan(
             readIndexMode: "unsupported",
@@ -158,7 +158,7 @@ public sealed class PlanCommandTests
                     hit: false,
                     fallbackReason: "readIndex disabled by mode."),
                 PlanToken: null))));
-        var command = new PlanCommand(service, preflightService, new StubRequestInputReader(RequestInputReadResult.Success(DefaultRequestJson)));
+        var command = new PlanCommand(service, preflightService, new StubRequestInputReader(RequestInputReadResult.Success(DefaultRequestJson)), CommandResultTestWriter.Create());
 
         var (exitCode, standardOutput) = await StandardOutputCapture.Execute(() => command.Plan(
             timeout: "abc",
