+++ tests/Ucli.Tests/Hosting/Cli/RefreshCommandTests.cs
@@ -39,7 +39,7 @@ public sealed class RefreshCommandTests
     public async Task Refresh_UsesRefreshServiceAndWritesCommandResult ()
     {
         var service = new StubRefreshService((_, _) => ValueTask.FromResult(SuccessResult));
-        var command = new RefreshCommand(service);
+        var command = new RefreshCommand(service, CommandResultTestWriter.Create());
         using var cancellationTokenSource = new CancellationTokenSource();
 
         var (exitCode, standardOutput) = await StandardOutputCapture.Execute(() => command.Refresh(
@@ -103,7 +103,7 @@ public sealed class RefreshCommandTests
             Errors: [],
             Outcome: ApplicationOutcome.Success,
             ReadPostcondition: readPostcondition)));
-        var command = new RefreshCommand(service);
+        var command = new RefreshCommand(service, CommandResultTestWriter.Create());
 
         var (exitCode, standardOutput) = await StandardOutputCapture.Execute(() => command.Refresh(
             projectPath: "/repo/UnityProject",
@@ -124,7 +124,7 @@ public sealed class RefreshCommandTests
     public async Task Refresh_WhenTimeoutIsInvalid_ReturnsInvalidArgumentWithoutCallingService ()
     {
         var service = new StubRefreshService((_, _) => throw new InvalidOperationException("Service should not be called."));
-        var command = new RefreshCommand(service);
+        var command = new RefreshCommand(service, CommandResultTestWriter.Create());
 
         var (exitCode, standardOutput) = await StandardOutputCapture.Execute(() => command.Refresh(
             timeout: "abc",
@@ -147,7 +147,7 @@ public sealed class RefreshCommandTests
     public async Task Refresh_WhenModeIsInvalid_ReturnsInvalidArgumentWithoutCallingService ()
     {
         var service = new StubRefreshService((_, _) => throw new InvalidOperationException("Service should not be called."));
-        var command = new RefreshCommand(service);
+        var command = new RefreshCommand(service, CommandResultTestWriter.Create());
 
         var (exitCode, standardOutput) = await StandardOutputCapture.Execute(() => command.Refresh(
             mode: "unsupported",
