+++ tests/Ucli.Tests/Hosting/Cli/StatusCommandTests.cs
@@ -27,7 +27,7 @@ public sealed class StatusCommandTests
                 CanAcceptExecutionRequests: false,
                 Runtime: null,
                 TimeoutMilliseconds: 1234))));
-        var command = new StatusCommand(service);
+        var command = new StatusCommand(service, CommandResultTestWriter.Create());
         using var cancellationTokenSource = new CancellationTokenSource();
 
         await StandardOutputCapture.Execute(() => command.Status(
@@ -46,7 +46,7 @@ public sealed class StatusCommandTests
     public async Task Status_WhenTimeoutIsInvalid_ReturnsInvalidArgumentWithoutCallingService ()
     {
         var service = new StubStatusService((_, _) => throw new InvalidOperationException("Service should not be called."));
-        var command = new StatusCommand(service);
+        var command = new StatusCommand(service, CommandResultTestWriter.Create());
 
         var (exitCode, standardOutput) = await StandardOutputCapture.Execute(() => command.Status(
             timeout: "abc",
